+++ databento/live/protocol.py
@@ -82,7 +82,7 @@ class DatabentoLiveProtocol(asyncio.BufferedProtocol):
         self._heartbeat_interval_s = heartbeat_interval_s
 
         self._dbn_decoder = databento_dbn.DBNDecoder(
-            upgrade_policy=VersionUpgradePolicy.UPGRADE,
+            upgrade_policy=VersionUpgradePolicy.UPGRADE_TO_V2,
         )
         self._gateway_decoder = GatewayDecoder()